Update FastIO.java and FastIODemo.java
After the test on the OJ platform SUSTechACM, I find that if we do the flush operation after all the output materials have been written into the buffered area, the running time will be saved about 66%. So I removed all the flush() in the println(), printf(), print() in the FastWriter, and added it into close(). 
Therefore, if you need to use this FastIO, then do not forget to write fastWriter.close()!